requirements.txt
langchain-core==0.3.*
langchain-community==0.3.*
langchain==0.3.*
langchain-openai==0.2.*
pydantic==2.*
python-dotenv==1.*
chromadb==0.5.*
tiktoken==0.7.*
fastapi==0.115.*
uvicorn==0.32.*
langchain-google-genai
google-generativeai